Minor adjustments, fixed map and start route button location, removed websocket info (temporary change) and changed vendor icon. Adjusted folders and imports.
