+++ src/components/user/UserReviewsFeed.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import VendorRatings from '../Reviews';
+import VendorRatings from './Reviews';
 
 export default function UserReviewsFeed() {
 	return (
