+++ src/components/user/UserSignup.jsx
@@ -1,42 +1,32 @@
 import React, { useState } from 'react';
 import { Form, Button, Card } from 'react-bootstrap';
-import { Navigate, Link } from 'react-router-dom';
-import {
-	doSignInUserWithEmailAndPassword,
-	doSignInWithGoogle,
-} from '../firebase/auth';
-import { useAuth } from '../contexts/authContext/index';
+import { useAuth } from '../../contexts/authContext';
+import { Navigate, Link, useNavigate } from 'react-router-dom';
+import { doCreateUserWithEmailAndPassword } from '../../firebase/auth';
 
-const UserLogin = () => {
-	const { userLoggedIn } = useAuth();
+const UserSignup = () => {
+	const navigate = useNavigate();
 
 	const [email, setEmail] = useState('');
 	const [password, setPassword] = useState('');
-	const [isSigningIn, setIsSigningIn] = useState(false);
+	const [confirmPassword, setconfirmPassword] = useState('');
+	const [isRegistering, setIsRegistering] = useState(false);
 	const [errorMessage, setErrorMessage] = useState('');
+
+	const { userLoggedIn } = useAuth();
 
 	const handleSubmit = async (e) => {
 		e.preventDefault();
-		if (!isSigningIn) {
-			setIsSigningIn(true);
-			await doSignInUserWithEmailAndPassword(email, password);
-		}
-	};
-
-	const onGoogleSignIn = (e) => {
-		e.preventDefault();
-		if (!isSigningIn) {
-			setIsSigningIn(true);
-			doSignInWithGoogle().catch((err) => {
-				setIsSigningIn(false);
-			});
+		if (!isRegistering) {
+			setIsRegistering(true);
+			await doCreateUserWithEmailAndPassword(email, password);
 		}
 	};
 
 	return (
 		<>
 			<div className="w-100" style={{ maxWidth: '400px' }}>
-				{userLoggedIn && <Navigate to={'/usermapview'} replace={true} />}
+				{userLoggedIn && <Navigate to={'/userdashboard'} replace={true} />}
 				<Link to="/">
 					<button type="button"
 						class="btn-close "
@@ -48,15 +38,16 @@ const UserLogin = () => {
 							borderRadius: '15px',
 							padding: '8px',
 							marginBottom: '5px',
-						}}
-					></button>
+						}}></button>
 				</Link>
+
 				<Card>
 					<Card.Body>
-						<h2 className="text-center mb-4">User Log In</h2>
+						<h2 className="text-center mb-4">User Sign Up</h2>
 						<Form onSubmit={handleSubmit}>
 							<Form.Group id="email">
 								<Form.Label>Email</Form.Label>
+
 								<Form.Control
 									type="email"
 									required
@@ -69,7 +60,9 @@ const UserLogin = () => {
 							<Form.Group id="password">
 								<Form.Label>Password</Form.Label>
 								<Form.Control
+									disabled={isRegistering}
 									type="password"
+									autoComplete="new-password"
 									required
 									value={password}
 									onChange={(e) => {
@@ -77,45 +70,43 @@ const UserLogin = () => {
 									}}
 								/>
 							</Form.Group>
-
+							<Form.Group id="password">
+								<Form.Label>Password Confirmation</Form.Label>
+								<Form.Control
+									disabled={isRegistering}
+									type="password"
+									autoComplete="off"
+									required
+									value={confirmPassword}
+									onChange={(e) => {
+										setconfirmPassword(e.target.value);
+									}}
+								/>
+							</Form.Group>
 							{errorMessage && (
 								<span className="text-red-600 font-bold">{errorMessage}</span>
 							)}
 							<Button
 								type="submit"
-								disabled={isSigningIn}
+								disabled={isRegistering}
 								style={{ backgroundColor: '#EA3187', borderColor: '#EA3187' }}
 								className={`mx-auto d-block text-white font-medium rounded-lg mt-2 ${
-									isSigningIn
+									isRegistering
 										? 'bg-gray-300 cursor-not-allowed'
 										: 'bg-indigo-600 hover:bg-indigo-700 hover:shadow-xl transition duration-300'
 								}`}
 							>
-								{isSigningIn ? 'Signing In...' : 'Sign In'}
+								{isRegistering ? 'Signing Up...' : 'Sign Up'}
 							</Button>
 						</Form>
 					</Card.Body>
 				</Card>
-				<Button
-					style={{ backgroundColor: '#EA3187', borderColor: '#EA3187' }}
-					disabled={isSigningIn}
-					onClick={(e) => {
-						onGoogleSignIn(e);
-					}}
-					className={`w-100 text-center mt-3 ${
-						isSigningIn
-							? 'cursor-not-allowed'
-							: 'hover:bg-gray-100 transition duration-300 active:bg-gray-100'
-					}`}
-				>
-					<div>{isSigningIn ? 'Signing In...' : 'Sign in with Google'}</div>
-				</Button>
-				<div className="w-100 text-center mt-5">
-					Need an account? <Link to="/usersignup">Sign Up</Link>
+				<div className="w-100 text-center mt-2">
+					Already have an account? <Link to="/userlogin">Log In</Link>
 				</div>
 			</div>
 		</>
 	);
 };
 
-export default UserLogin;
+export default UserSignup;
