--- src/styling.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" type="text/css" href="styling.css">
-    <title>Where's My Ice Cream</title>
-</head>
-<body>
-    <div class="navigation">
-        <ul>
-            <li class = "list active">
-                <a href="#">
-                    <span class="icon">
-                    <ion-icon name="home-outline"></ion-icon>
-                </span>
-                    <span class="text">Home</span>
-                </a>
-            </li>
-            <li class = "list">
-                <a href="#">
-                    <span class="icon">
-                        <ion-icon name="calendar-outline"></ion-icon>
-                    </span>
-                    <span class="text">Catering</span>
-                </a>
-            </li>
-            <li class = "list">
-                <a href="#">
-                    <span class="icon">
-                        <ion-icon name="star-outline"></ion-icon>
-                    </span>
-                    <span class="text">Ratings</span>
-                </a>
-            </li>
-            <div class="indicator"></div>
-        </ul>
-    </div>
-
-    <script>
-        const list = document.querySelectorAll('.list');
-        function activeLink(){
-            list.forEach((item) => item.classList.remove('active'));
-            this.classList.add('active');
-        }
-        list.forEach((item) =>
-        item.addEventListener('click', activeLink));
-    </script>
-    <script type="module" src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"></script>
-<script nomodule src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.js"></script>
-<style>
-    @import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:wght@300&display=swap')
-    </style>
-</body>
-</html>
